+++ GTiff2Tiles.Core/Enums/Image/Interpolations.cs
@@ -3,11 +3,11 @@
     /// <summary>
     /// Represents the interpolation algorithms.
     /// </summary>
-    public static class Interpolations
+    internal static class Interpolations
     {
         /// <summary>
         /// Bicubic interpoilation.
         /// </summary>
-        public const string Bicubic = "bicubic";
+        internal const string Bicubic = "bicubic";
     }
 }