+++ GTiff2Tiles.Core/Helpers/NetVipsHelper.cs
@@ -5,13 +5,13 @@ namespace GTiff2Tiles.Core.Helpers
     /// <summary>
     /// Some methods for initializing NetVips.
     /// </summary>
-    public static class NetVipsHelper
+    internal static class NetVipsHelper
     {
         //todo initialize NetVips location.
 
         /// <summary>
         /// Disables NetVips log warnings.
         /// </summary>
-        public static void DisableLog() => Log.SetLogHandler("VIPS", NetVips.Enums.LogLevelFlags.Warning, (domain, level, message) => { });
+        internal static void DisableLog() => Log.SetLogHandler("VIPS", NetVips.Enums.LogLevelFlags.Warning, (domain, level, message) => { });
     }
 }